feat: document the client controller API (#11)
